requirements.txt
absl-py==0.8.1
alabaster==0.7.12
altgraph==0.17
asgiref==3.2.10
asn1crypto==1.2.0
astor==0.8.0
astroid==2.3.3
astunparse==1.6.3
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.9.1
django-filter==2.3.0
django-forms-bootstrap==3.1.0
django-jquery==3.1.0
django-static-jquery==2.1.4
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djongo==1.3.3
docutils==0.15.2
easydict==1.9
entrypoints==0.3
Flask==1.1.1
future==0.18.2
gast==0.3.3
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
keyring==19.2.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
mysqlclient==2.0.1
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.1
numba==0.48.0
numpy==1.18.1
numpydoc==0.9.1
oauthlib==3.1.0
object-detection==0.1
olefile==0.46
opencv-contrib-python==4.2.0.32
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==5.2.0
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.0
psutil==5.6.5
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyInstaller==3.6
pylint==2.4.4
pymongo==3.10.1
pyodbc==4.0.27
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.6
spyder-kernels==0.5.2
sqlparse==0.2.4
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==0.4.1
torchvision==0.2.1
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
zipp==0.6.0
